Added Salvage
Added Salvage prefab
Spawning is identical to scrap behavior
Salvage has its own tier/data scriptable objects
When collided with player, the CollectableManager stores the data of that salvage object in a list to be processed later.

Scrap now does not add itself to the money counter immediately after being collected
The total value of all scrap collected is stored in the CollectableManager and only when the player returns to the HUB is that total added to the money counter
That total is then reset back to 0

+++ Transluminal/Assets/Scripts/ScriptableObjects/ScrapData.cs
@@ -6,5 +6,5 @@ public class ScrapData : ScriptableObject
 {
     public TierData tier;
     public Sprite sprite;
-    public int scale;
+    public float scale;
 }
